+++ src/hooks/use-theme-series-colors.ts
@@ -1,11 +1,15 @@
 import {useMemo} from 'react';
 
-import type {Theme, Color, LineStyle, DataSeries} from '../types';
+import type {Theme, Color, LineStyle, DataPoint, Data} from '../types';
 
-// We don't use the data property anywhere, so it's
-// safe to use "any" here as the type passed to DataSeries.
-// We really only care about color and lineStyle
-interface ValidData extends DataSeries<any, Color> {
+// Note: This is a bandaid until all component are using
+// the new DataSeries type.
+interface ValidData {
+  data: (DataPoint | Data)[];
+  color?: Color;
+  isComparison?: boolean;
+  name?: string;
+
   lineStyle?: LineStyle;
 }
 
@@ -14,7 +18,7 @@ function getFilteredSeriesCount(series: Partial<ValidData>[]): number {
   // count when grabbing the series color.
   return (
     series.filter((item) => {
-      if (!item.lineStyle) {
+      if (item.isComparison !== true && !item.lineStyle) {
         return true;
       }
       return item.lineStyle === 'solid';